Rewrote isTouching, reverted ints back to floats. Cube sometimes jumps down out of the map.

+++ Game/Obstacle.java
@@ -1,13 +1,13 @@
 class Obstacle {
-  private int xcor, ycor;
-  public Obstacle(int x, int y) {
+  private float xcor, ycor;
+  public Obstacle(float x, float y) {
     xcor=x;
     ycor=y;
   }
-  public int getX() {
+  public float getX() {
     return xcor;
   }
-  public int getY() {
+  public float getY() {
     return ycor;
   }
   public void display() {
